Fix: Changed parameter type from Float to Double in FilterProductsDTO

+++ webshop-service/src/main/java/rs/ac/uns/ftn/webshopservice/dto/request/FilterProductsDTO.java
@@ -11,7 +11,7 @@ public class FilterProductsDTO {
 
     private String name;
     private String categoryName;
-    private Float downPrice;
-    private Float topPrice;
+    private Double downPrice;
+    private Double topPrice;
     private Boolean isFreeShipping;
 }
